Share the coroutine channel used by resource pools
Move the duplicate object and connection pool channels into Coroutine\PoolChannel. Keep the resource queue separate from native wake signals so queued resources remain available for draining after close.

Consolidate channel tests while retaining FIFO behavior, cancellation classification, waiter cleanup, signal coalescing and coroutine/non-coroutine transitions.
